+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New Management endpoints for uploading multiple files (limited to 2GB):
 - `/api/DataStaging/UploadFilesToProjectDir` - uploads files to project directory
 - `/api/DataStaging/UploadJobScriptsToProjectDir` - upload job scripts to project directory and makes them executable
-- `/api/DataStaging/UploadFilesToJobExecutionDir` - uploadfiles to job execution directory
+- `/heappe/FileTransfer/UploadFilesToJobExecutionDir` - uploadfiles to job execution directory
 
 ## V6.1.1